Makes the helper to check the exception message local
